Values validation fix

+++ server/web/routes/home_page.py
@@ -78,22 +78,31 @@ def create_new_device():
     """
     payload = request.json
     payload["errors"] = []
-    device_id = payload["deviceData"]["device_id"]
-    if not server.get_device_by_device_id(device_id):
-        device = Devices()
-        device.device_id = device_id
-        device.device_type_id = payload["deviceData"]["device_type"]
-        device.device_access_key = payload["deviceData"]["device_access_key"]
-        database.session.add(device)
-        database.session.commit()
 
-        server.update_allowed_devices()
-        return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CREATED.value)
+    device_id = payload["deviceData"]["device_id"]
+    device_type_id = payload["deviceData"]["device_type"]
+    device_access_key = payload["deviceData"]["device_access_key"]
+
+    if device_id and device_type_id and device_access_key:
+        if not server.get_device_by_device_id(device_id):
+            device = Devices()
+            device.device_id = device_id
+            device.device_type_id = device_type_id
+            device.device_access_key = device_access_key
+            database.session.add(device)
+            database.session.commit()
+
+            server.update_allowed_devices()
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CREATED.value)
+        else:
+            payload["errors"] = ["Device with this ID already exists"]
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CONFLICT.value)
     else:
-        payload["errors"] = ["Device with this ID already exists"]
+        payload["errors"] = ["One or multiple required fields were empty"]
         return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CONFLICT.value)
+                                  status=ResponseStatus.PRECONDITION_FAILED.value)
 
 
 @process_rest_request   # Make sure this decorator goes before routing to prevent namespace_overwrite errors
@@ -118,21 +127,31 @@ def update_device_details():
     """
     payload = request.json
     payload["errors"] = []
-    device = Devices.query.get(int(payload["deviceData"]["id"]))
-    if device:
-        device.device_id = payload["deviceData"]["device_id"]
-        device.device_type_id = payload["deviceData"]["device_type"]
-        received_access_key = payload["deviceData"]["device_access_key"]
-        device.device_access_key = received_access_key if received_access_key else device.device_access_key
-        database.session.commit()
 
-        server.update_allowed_devices()
-        return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.OK.value)
+    db_id = payload["deviceData"]["id"]
+    device_id = payload["deviceData"]["device_id"]
+    device_type_id = payload["deviceData"]["device_type"]
+    device_access_key = payload["deviceData"]["device_access_key"]
+
+    if db_id and device_id and device_type_id:
+        device = Devices.query.get(int(db_id))
+        if device:
+            device.device_id = device_id
+            device.device_type_id = device_type_id
+            device.device_access_key = device_access_key if device_access_key else device.device_access_key
+            database.session.commit()
+
+            server.update_allowed_devices()
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.OK.value)
+        else:
+            payload["errors"].append("Device with specified ID doesn't exist")
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CONFLICT.value)
     else:
-        payload["errors"].append("Device with specified ID doesn't exist")
+        payload["errors"] = ["One or multiple required fields were empty"]
         return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CONFLICT.value)
+                                  status=ResponseStatus.PRECONDITION_FAILED.value)
 
 
 @process_rest_request   # Make sure this decorator goes before routing to prevent namespace_overwrite errors
@@ -157,21 +176,28 @@ def delete_device():
     payload = request.json
     payload["errors"] = []
     device_id = payload["deviceData"]["device_id"]
-    if server.get_device_by_device_id(device_id):
-        device = Devices.query.filter(Devices.device_id == device_id)
-        # Since system monitor logs are detached from Device model - we delete it first, separately
-        DeviceSystemMonitorReadings.query.filter(DeviceSystemMonitorReadings.device_id == device.first().id).delete()
-        # Once done - we delete device with it's included dependencies
-        device.delete()
-        database.session.commit()
 
-        server.update_allowed_devices()
-        return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.OK.value)
+    if device_id:
+        if server.get_device_by_device_id(device_id):
+            device = Devices.query.filter(Devices.device_id == device_id)
+            # Since system monitor logs are detached from Device model - we delete it first, separately
+            DeviceSystemMonitorReadings.query.filter(
+                DeviceSystemMonitorReadings.device_id == device.first().id).delete()
+            # Once done - we delete device with it's included dependencies
+            device.delete()
+            database.session.commit()
+
+            server.update_allowed_devices()
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.OK.value)
+        else:
+            payload["errors"] = ["Device with this ID doesn't exist."]
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.NOT_FOUND.value)
     else:
-        payload["errors"] = ["Device with this ID doesn't exist."]
+        payload["errors"] = ["One or multiple required fields were empty"]
         return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.NOT_FOUND.value)
+                                  status=ResponseStatus.PRECONDITION_FAILED.value)
 
 
 @web_service.route("/home/getUsersList", methods=["GET"])
@@ -243,29 +269,40 @@ def create_user():
     """
     payload = request.json
     payload["errors"] = []
-    user_email = payload["userData"]["email"]
-    user = Users.query.filter(Users.email == user_email).first()
-    if not user:
-        roles_to_add = []
-        for role_requested in payload["userData"]["roles"]:
-            role_name = role_requested["role_name"]
-            db_role_instance = user_datastore.find_role(role_name)
-            if db_role_instance:
-                roles_to_add.append(role_name)
-        user = user_datastore.create_user(email=user_email,
-                                          password=hash_password(payload["userData"]["password"]),
-                                          roles=roles_to_add)
-        user.first_name = payload["userData"]["first_name"]
-        user.last_name = payload["userData"]["last_name"]
-        user.active = bool(payload["userData"]["active"])
-        database.session.commit()
 
-        return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CREATED.value)
+    user_email = payload["userData"]["email"]
+    user_password = payload["userData"]["password"]
+    user_first_name = payload["userData"]["first_name"]
+    user_last_name = payload["userData"]["last_name"]
+    user_active = payload["userData"]["active"]
+
+    if user_email and user_password and user_first_name and user_last_name:
+        user = Users.query.filter(Users.email == user_email).first()
+        if not user:
+            roles_to_add = []
+            for role_requested in payload["userData"]["roles"]:
+                role_name = role_requested["role_name"]
+                db_role_instance = user_datastore.find_role(role_name)
+                if db_role_instance:
+                    roles_to_add.append(role_name)
+            user = user_datastore.create_user(email=user_email,
+                                              password=hash_password(user_password),
+                                              roles=roles_to_add)
+            user.first_name = user_first_name
+            user.last_name = user_last_name
+            user.active = bool(user_active)
+            database.session.commit()
+
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CREATED.value)
+        else:
+            payload["errors"] = ["User with this email already exists"]
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CONFLICT.value)
     else:
-        payload["errors"] = ["User with this email already exists"]
+        payload["errors"] = ["One or multiple required fields were empty"]
         return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CONFLICT.value)
+                                  status=ResponseStatus.PRECONDITION_FAILED.value)
 
 
 @process_rest_request   # Make sure this decorator goes before routing to prevent namespace_overwrite errors
@@ -296,34 +333,46 @@ def update_user():
     """
     payload = request.json
     payload["errors"] = []
-    user = Users.query.get(int(payload["userData"]["id"]))
-    if user:
-        requested_roles = [role["role_name"] for role in payload["userData"]["roles"]]
-        existing_roles = [role.name for role in user.roles]
-        if not requested_roles == existing_roles:
-            for existing_role in existing_roles:
-                user_datastore.remove_role_from_user(user.email, existing_role)
 
-            for requested_role in requested_roles:
-                user_datastore.add_role_to_user(user.email, requested_role)
+    user_id = payload["userData"]["id"]
+    user_email = payload["userData"]["email"]
+    user_password = payload["userData"]["password"]
+    user_first_name = payload["userData"]["first_name"]
+    user_last_name = payload["userData"]["last_name"]
+    user_active = payload["userData"]["active"]
 
-        requested_password = payload["userData"]["password"]
-        if requested_password:
-            user.password = hash_password(requested_password)
+    if user_id and user_email and user_first_name and user_last_name:
+        user = Users.query.get(int(user_id))
+        if user:
+            requested_roles = [role["role_name"] for role in payload["userData"]["roles"]]
+            existing_roles = [role.name for role in user.roles]
+            if not requested_roles == existing_roles:
+                for existing_role in existing_roles:
+                    user_datastore.remove_role_from_user(user.email, existing_role)
 
-        user.email = payload["userData"]["email"]
-        user.first_name = payload["userData"]["first_name"]
-        user.last_name = payload["userData"]["last_name"]
-        user.active = bool(payload["userData"]["active"])
+                for requested_role in requested_roles:
+                    user_datastore.add_role_to_user(user.email, requested_role)
 
-        database.session.commit()
+            if user_password:
+                user.password = hash_password(user_password)
 
-        return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.OK.value)
+            user.email = user_email
+            user.first_name = user_first_name
+            user.last_name = user_last_name
+            user.active = bool(user_active)
+
+            database.session.commit()
+
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.OK.value)
+        else:
+            payload["errors"] = ["User with this ID doesn't exist"]
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CONFLICT.value)
     else:
-        payload["errors"] = ["User with this ID doesn't exist"]
+        payload["errors"] = ["One or multiple required fields were empty"]
         return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CONFLICT.value)
+                                  status=ResponseStatus.PRECONDITION_FAILED.value)
 
 
 @process_rest_request   # Make sure this decorator goes before routing to prevent namespace_overwrite errors
@@ -348,15 +397,20 @@ def delete_user():
     payload = request.json
     payload["errors"] = []
     user_email = payload["userData"]["email"]
-    user = Users.query.filter(Users.email == user_email).first()
-    if user:
-        user_datastore.delete_user(user)
+    if user_email:
+        user = Users.query.filter(Users.email == user_email).first()
+        if user:
+            user_datastore.delete_user(user)
 
-        database.session.commit()
+            database.session.commit()
 
-        return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.OK.value)
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.OK.value)
+        else:
+            payload["errors"] = ["User with this email doesn't exist"]
+            return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
+                                      status=ResponseStatus.CONFLICT.value)
     else:
-        payload["errors"] = ["User with this email doesn't exist"]
+        payload["errors"] = ["One or multiple required fields were empty"]
         return utils.get_response(payload, mimetype=MimeType.JSON_MIMETYPE.value,
-                                  status=ResponseStatus.CONFLICT.value)
+                                  status=ResponseStatus.PRECONDITION_FAILED.value)
